requirements.txt
asgiref==3.6.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
coreapi==2.3.3
coreschema==0.0.4
cymem==2.0.7
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
docx2txt==0.8
drf-yasg==1.21.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.2
murmurhash==1.0.9
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
pdfminer==20191125
preshed==3.0.8
pycryptodome==3.17
pydantic==1.10.4
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
smart-open==6.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.5
thinc==8.1.7
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
wasabi==1.1.1
